backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.3
opencv-python==4.9.0.80
PyYAML==6.0.1
Werkzeug==3.0.1
